Implement incoming call event + Change the voice manager to not create voice channels but to use the real channel id #1

+++ server/session.cpp
@@ -77,7 +77,7 @@ void Session::recv()
 																}
 
 																// Set the size of the response
-																header.size = response.size();
+																header.size = (unsigned int)response.size();
 
 																// Encode the header in a string
 																header_str = Utils::EncodeHeader(header);
