+++ shared/packets/voice_packet.h
@@ -41,7 +41,7 @@ public:
 			_channel_id = params[1];
 			
 			// Parse voice data
-			_voice_data_len = params[2].length();
+			_voice_data_len = (unsigned int)params[2].length();
 			_voice_data = new char[_voice_data_len];
 			memcpy(_voice_data, params[2].c_str(), _voice_data_len);
 		} else {
